2nd week test q2

+++ 2nd weel test/question 1 box cost.cpp
@@ -91,7 +91,7 @@ class Box{			//box
 			this->height = height;
 			this->part = part;
 		}
-		int calculateCost(Flute flute){
+		long long int calculateCost(Flute flute){
 			area = length*height*width;
 			totalGSM = (flute.extraGSM+flute.GSM)*100;
